backend/requirements.txt
# Inbox Copilot Backend Dependencies

# FastAPI framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Google API
google-api-python-client==2.111.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# HTTP client
httpx==0.26.0

# Settings management
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
